--- a/lib/admin_user_index_query.rb
+++ b/lib/admin_user_index_query.rb
@@ -93,7 +93,7 @@ class AdminUserIndexQuery
 
   def filter_by_search
     if params[:email].present?
-      return @query.where('user_emails.email = ?', params[:email].downcase)
+      return @query.joins(:primary_email).where('user_emails.email = ?', params[:email].downcase)
     end
 
     filter = params[:filter]